// frontend/src/app/core/api/models/read-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { IdNameDto } from '../models/id-name-dto';
export interface ReadUserDto {
  coop?: IdNameDto;
  email: string;
  gender: 'm' | 'f';
  id: string;
  name: string;
  role: 'admin' | 'coop_admin' | 'user';
  yearOfBirth: number;
}